Update README to include Babel example
I use ES2015 in most of my JavaScript projects now and I suspect others are in the same situation. The ES2015 syntax additions may be even more popular than CoffeeScript at the moment. Since there is a CoffeeScript example, I think it would be helpful to include a brief Babel example as well, just to help people get on the right path.

Since Babel 6 does not automatically transpile ES2015, I also added a link to the setup instructions. I'd like to link directly to the Mocha instructions, but that doesn't seem to be possible.